获取视频第一帧图片

+++ MyApp/Media.h
@@ -10,8 +10,8 @@
 
 @interface Media : NSObject
 @property (nonatomic, assign) NSInteger media_id;
-@property (nonatomic, assign) int media_width;
-@property (nonatomic, assign) int media_height;
+@property (nonatomic, assign) CGFloat media_width;
+@property (nonatomic, assign) CGFloat media_height;
 @property (nonatomic, copy) NSString *media_url;
 @property (nonatomic, assign) short media_type;
 @end
